refactor: separate HTML element constants into a dedicated module
- Moved HTML element constants from Global.ts to a new module Global_HTMLElements.ts
- Updated imports in relevant files to use the new Global_HTMLElements.ts module
- Improved code organization and readability
