+++ src/Modules/Global.ts
@@ -9,14 +9,6 @@ import {
   IPlayerResultHandler
 } from './Global.d'
 
-export const GameWinnerHolder = document.getElementById('gameWinnerHolder') as HTMLElement;
-export const GameLossHolder = document.getElementById('gameLossHolder') as HTMLElement;
-export const ComputerNameHolder = document.getElementById('computerName') as HTMLElement | null;
-export const PlayerNameHolder = document.getElementById('playerName') as HTMLElement | null;
-export const RollAnnouncePlayer = document.getElementById('currentPlayerName') as HTMLElement | null;
-export const RollAnnounceAction = document.getElementById('currentPlayerAction') as HTMLElement | null;
-export const ScoreBoardHolder = document.getElementById('scoreBoardHolder') as HTMLElement | null;
-
 const playerButtons = document.querySelectorAll('.playerColumn button');
 const winnerNameHolder = document.getElementById('winnerName');
 const roundOrGameStrHolder = document.getElementById('roundOrGame');
