+++ src/Modules/Global_HTMLElements.ts
@@ -1,0 +1,7 @@
+export const GameWinnerHolder = document.getElementById('gameWinnerHolder') as HTMLElement;
+export const GameLossHolder = document.getElementById('gameLossHolder') as HTMLElement;
+export const ComputerNameHolder = document.getElementById('computerName') as HTMLElement | null;
+export const PlayerNameHolder = document.getElementById('playerName') as HTMLElement | null;
+export const RollAnnouncePlayer = document.getElementById('currentPlayerName') as HTMLElement | null;
+export const RollAnnounceAction = document.getElementById('currentPlayerAction') as HTMLElement | null;
+export const ScoreBoardHolder = document.getElementById('scoreBoardHolder') as HTMLElement | null;
